delete result.py and edit retrieval.py
